# fastfloat: support elided integer or float part